An attempt to resolve https://github.com/UAVCAN/specification/pull/46/files/fffe93b4a9dc1f3cd08a8687eabcb3caa28b5e95#r277125281
I have somewhat strictened the port-ID assignment rules to avoid possible confusion; if necessary, we can go back to
less strict rules in the next minor version of the specification since the backward compatibility would not be compromised.

There is some minor duplication of text in chapters 2 and 3; this is probably acceptable because it helps us to reduce the
number of forward references.