CAN ID layout update
Updated the CAN ID layout to accommodate the v0/v1 compatibility objectives outlined in this thread:
https://forum.uavcan.org/t/uavcan-v1-0-and-ardupilot/671.

The updated format keeps the source node-ID at the same offset regardless of the protocol version.
This allows v0/v1 nodes to share the same bus with no risk of CAN ID collision.

The change required me to move the version bit from the 0th position to the 23rd position.
The new position is chosen to allow easy expansion of the service-ID field shall that become necessary in the future.
Also, for message transfers, this bit was already effectively reserved, even though it has not been explicitly marked as such,
because the range of the 16-bit subject-ID field was limited to [0, 32767], meaning that the most significant bit,
which is 23rd, was to be always zero.

The message transfer format got an extra reserved bit at the 7th position to keep the subject-ID aligned.