Attribution of the CAN acceptance filter configuration algorithm.
The paper "Analysis and Optimization of Message Acceptance Filter Configurations for Controller Area Network (CAN)" published in 2017
reported that there was no prior work on the subject. Technically, this is not true, although the mistake is understandable because
our approach was never published in any scientific databases. The explicit attribution may reduce the chances of this happening in the future.