Add a remark about fixing @extent when the bit length set is changed.
Scott: Also needed is a discussion of the effect of removing a field from a datatype that uses this default extent. By shrinking the serializable size of the type I'd also be shrinking the extent which seems problematic, right?

Pavel: Yes, but I think this is covered adequately in 3.8.3.2, in the beginning [...]. Do you think this needs to be expanded?

Scott: Perhaps we remind the user that, if they update a datatype by adding a field, they should also consider fixing the extent at the previously defined value.